+++ client/src/app/core/services/openslides-status.service.spec.ts
@@ -1,0 +1,17 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { E2EImportsModule } from 'e2e-imports.module';
+import { OpenSlidesStatusService } from './openslides-status.service';
+
+describe('OpenSlidesStatusService', () => {
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            imports: [E2EImportsModule],
+            providers: [OpenSlidesStatusService]
+        });
+    });
+
+    it('should be created', inject([OpenSlidesStatusService], (service: OpenSlidesStatusService) => {
+        expect(service).toBeTruthy();
+    }));
+});