+++ client/src/app/site/mediafiles/services/mediafile-repository.service.ts
@@ -78,7 +78,7 @@ export class MediafileRepositoryService extends BaseRepository<ViewMediafile, Me
     public async uploadFile(file: FormData): Promise<Identifiable> {
         const restPath = `rest/mediafiles/mediafile/`;
         const emptyHeader = new HttpHeaders();
-        return this.httpService.post<Identifiable>(restPath, file, emptyHeader);
+        return this.httpService.post<Identifiable>(restPath, file, {}, emptyHeader);
     }
 
     /**