+++ client/src/app/site/users/services/user-repository.service.ts
@@ -21,7 +21,11 @@ import { TranslateService } from '@ngx-translate/core';
 })
 export class UserRepositoryService extends BaseRepository<ViewUser, User> {
     /**
-     * Constructor calls the parent constructor
+     * Constructor for the user repo
+     *
+     * @param DS The DataStore
+     * @param mapperService Maps collection strings to classes
+     * @param dataSend sending changed objects
      */
     public constructor(
         DS: DataStoreService,